+++ components.d.ts
@@ -8,7 +8,6 @@ export {}
 declare module '@vue/runtime-core' {
   export interface GlobalComponents {
     Editor: typeof import('./src/components/editor/Editor.vue')['default']
-    ElIcon: typeof import('element-plus/es')['ElIcon']
     PublishContainer: typeof import('./src/components/editor/PublishContainer.vue')['default']
     RouterLink: typeof import('vue-router')['RouterLink']
     RouterView: typeof import('vue-router')['RouterView']
