+++ electron/extensions/juejin/tsconfig.json
@@ -100,5 +100,5 @@
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true /* Skip type checking all .d.ts files. */
   },
-  "include": ["../../../src/dew-dts/dew.d.ts"]
+  "include": ["../../../dew-dts/dew.d.ts"]
 }
