+++ electron/main/services/extensionHostConnection.ts
@@ -1,18 +1,12 @@
-import type { ChildProcess } from "child_process";
+import type { ChildProcess } from 'child_process'
 import * as path from 'path'
 import { fork } from 'child_process'
 import { ProcessMessage } from './processMessage.impl'
 
-import { ProcessMessage } from './processMessage.impl'
-
-
 export class ExtensionConnection {
   private static _instance: ExtensionConnection
 
-  private static _instance: ExtensionConnection
-
   private extensionProcess: ChildProcess
-  private constructor() { }
 
   public static getInstance() {
     if (!ExtensionConnection._instance) {
@@ -20,19 +14,11 @@ export class ExtensionConnection {
     }
     return ExtensionConnection._instance
   }
-  private constructor() { }
-
-  public static getInstance() {
-    if (!ExtensionConnection._instance) {
-      ExtensionConnection._instance = new ExtensionConnection()
-    }
-    return ExtensionConnection._instance
-  }
+  private constructor() {}
 
   public async start(startParams?: any) {
     this.extensionProcess = fork(path.resolve(__dirname, './extensionProcessMain.js'))
 
-    this.extensionProcess.on('message', (message: ProcessMessage) => {
     this.extensionProcess.on('message', (message: ProcessMessage) => {
       console.log('main process')
       console.log(message)
@@ -41,9 +27,9 @@ export class ExtensionConnection {
     setTimeout(() => {
       this.extensionProcess.send({
         type: 'plugin:start',
-        data: null
+        data: null,
       })
-    }, 1000 * 4);
+    }, 1000 * 4)
   }
 
   public restartProcess() {
@@ -51,17 +37,13 @@ export class ExtensionConnection {
 
     setTimeout(() => {
       this.start()
-    }, 1000);
+    }, 1000)
   }
 
   public sendMsg(type: string, data: any) {
     this.extensionProcess.send({
       type,
-      data
+      data,
     })
   }
-
-
-
-
 }
