+++ electron/preload/index.ts
@@ -1,8 +1,8 @@
 const { contextBridge, ipcRenderer } = require('electron')
 
 contextBridge.exposeInMainWorld('electronAPI', {
-  restartProcess: (arg) => ipcRenderer.invoke('blog.restart', arg),
+  restartProcess: (arg) => ipcRenderer.invoke('blog:restart', arg),
   createArticle: (arg) => ipcRenderer.invoke('blog:create', arg),
   updateArticle: (arg) => ipcRenderer.invoke('blog:update', arg),
-  publishArticle: (arg: any) => ipcRenderer.invoke('blog:publish', arg)
+  publishArticle: (arg: any) => ipcRenderer.invoke('blog:publish', arg),
 })
